Dave Perry:
I have modeled N7764P, the comanche 250 I co-own with 2 other Seagate engineers.
In the process of doing this model, I have also improved, worked on, or added
instruments, etc. to Instruments-3d.

* I fixed the adf so the azimuth card is tied to the correct property
* Added TO, From, and out-of range indicators to the 3d vor.
* Also added a manifold pressure gage and a pa24-250 asi (using digital photos
  of the actual asi as the starting textures.

Affects on other AC.  As far as I know, this will add to-from to the vor and
correct the adf behavior for pa28-161.  It would be a trivial edit to the
pa28-161.xml to give the same vor performance to that AC.
Now both VORs act as one.