Changes from Dave Perry:
I made a minor change to one file in the c172p that accommodates a
feature I added to the navcom-kx155.xml and the dme.xml in the
Instruments folder.  I wanted to leave the frequencies, etc. on these
dark until there was voltage applied to
/systems/electrical/outputs/nav.  This was accomplished by adding a
param and property alias  pointing to the appropriate value.  Since the
c172p "always" applies 28 volts, this made no change in the c172p, but
allowed me to model the avionics master and battery master realisticly
in pa24-electrical.nas.  I put in comments explaining this at the change
points.

There is a useful  "help > aircraft help" that walks you through the
start procedure and lists the key bindings as well as the key to "light
up" the hot spots.